requirements.txt
requests==2.22.0
discord.py==1.3.2